do not depend on undefined behavior
llvm-svn: 31946

+++ SingleSource/UnitTests/FloatPrecision.c
@@ -2,7 +2,7 @@
 
 void print(float X) {
   printf("%f\n", X*2147483647);
-  printf("%x\n", (unsigned)(X*2147483647));
+  printf("%x\n", (int)(unsigned long long)(X*2147483647));
 }
 
 int main() {
